Modification on code as per Ashish Sir code review

+++ BikeWale.UI/controls/ExpertReviewsVideos.ascx.cs
@@ -16,7 +16,7 @@ namespace Bikewale.Controls
         public EnumVideosCategory CategoryId { get; set; }
         public ushort TotalRecords { get; set; }
         public string SectionTitle { get; set; }
-        public int FetchedRecordsCount { get; set; }
+        public ushort FetchedRecordsCount { get; set; }
         public string SectionBackgroundClass { get; set; }
         protected BikeVideoEntity FirstVideoRecord { get; set; }
         public string CategoryIdList { get; set; }
